+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     - Fix a target sequence extracting bug.
     - Add output of query start and end positions.
     - Add a new flag `-w/--load-whole-seeds` for loading the whole seed data into memory for faster search.
+    - Keep index of genome data in memory.
 - `lexicmap utils kmers`:
     - Fix a little bug, wrong number of k-mers for the second k-mer in each k-mer pair.
 - New commands: